fix(i18n): address PR review — unify gateway language policy, move side-effect to services
Route gateway's language resolution through resolveActiveLanguage so the
doc comment's "one place" claim is accurate. Move reconcileInterfaceLanguage
from utils/ (pure helpers) to services/i18n.ts (side-effect singletons)
per apps/web AGENTS.md.
