+++ apps/gateway/src/routers/root.router.ts
@@ -1,4 +1,4 @@
-import { $Language } from '@opendatacapture/schemas/core';
+import { $Language, resolveActiveLanguage } from '@opendatacapture/schemas/core';
 import { $InstrumentBundleContainer } from '@opendatacapture/schemas/instrument';
 import { Router } from 'express';
 
@@ -45,10 +45,9 @@ router.get(
     // resolve the same language; anything else renders the page in English and then swaps it.
     const { activeLanguages } = assignment;
     const requestedLanguage = $Language.safeParse(req.query.lang);
-    const language =
-      requestedLanguage.success && activeLanguages.includes(requestedLanguage.data)
-        ? requestedLanguage.data
-        : activeLanguages[0];
+    const language = requestedLanguage.success
+      ? resolveActiveLanguage(requestedLanguage.data, activeLanguages)
+      : activeLanguages[0];
 
     const token = generateToken(assignment.id);
     const html = res.locals.loadRoot({
